NTP-1008: Refactor the query to get the enquirer to view all responses faster.
