+++ Application/Common/Interfaces/Repositories/ITuitionPartnerEnquiryRepository.cs
@@ -5,7 +5,5 @@ namespace Application.Common.Interfaces.Repositories;
 
 public interface ITuitionPartnerEnquiryRepository : IGenericRepository<TuitionPartnerEnquiry>
 {
-    Task<EnquirerViewAllResponsesModel> GetEnquirerViewAllResponses(int enquiryId, string baseServiceUrl);
-
     Task<EnquirerViewResponseModel?> GetEnquirerViewResponse(int enquiryId, int tuitionPartnerId);
 }
